% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ead-data.R
\docType{data}
\name{ead_data}
\alias{ead_data}
\title{URL of environment data}
\format{
A vector with 1 value:
\describe{
  \item{1}{url of environment data}
}
}
\source{
\url{http://www.environment.data.gov.uk/}
}
\usage{
ead_data
}
\description{
A dataset containing values for access the API
}
\keyword{datasets}
