% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-observations.R
\name{get_observations}
\alias{get_observations}
\title{Get Observations}
\usage{
get_observations(
  site_id = NULL,
  take = 500,
  skip = 0,
  type_id = NULL,
  property_id = NULL,
  ultimate_foi_id = NULL,
  date_from = NULL,
  date_to = NULL
)
}
\arguments{
\item{site_id}{list of sites ids.}

\item{take}{integer max number of returned observations}

\item{skip}{integer number of response to skip}

\item{type_id}{Type of observation}

\item{property_id}{Property ID}

\item{ultimate_foi_id}{Particular feature of interest}

\item{date_from}{Date string '2020-11-01'}

\item{date_to}{Date string '2021-02-01'}
}
\value{
dataframe
}
\description{
Query observations end point.
}
\examples{
site_id = c("/environment.data.gov.uk/ecology/site/bio/43378",
     "/environment.data.gov.uk/ecology/site/bio/10254")
obs <- get_observations(site_id)
}
