% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-properties.R
\name{get_properties}
\alias{get_properties}
\title{Get property info}
\usage{
get_properties()
}
\value{
Data frame
}
\description{
To get a full list of all Observable Properties in Observations, you can use
the /properties endpoint - this will return all observable properties across
all source datasets.
}
\examples{
properties <- get_properties()
}
