% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-site-info.R
\name{get_site_info}
\alias{get_site_info}
\title{Get site info}
\usage{
get_site_info(site_id = NULL)
}
\arguments{
\item{site_id}{List of site identifier}
}
\value{
Data frame
}
\description{
Biosys site information (such as altitude, slope etc).
}
\examples{
site_info <-
get_site_info(site_id = c("http://environment.data.gov.uk/ecology/site/bio/43378",
"http://environment.data.gov.uk/ecology/site/fish/10060"))
}
