% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sites.R
\name{get_sites}
\alias{get_sites}
\title{Get sites}
\usage{
get_sites(
  skip = 0,
  take = 250,
  mode = "minimal",
  type = NULL,
  lat = NULL,
  long = NULL,
  distance = NULL
)
}
\arguments{
\item{skip}{integer to skip}

\item{take}{integer to take}

\item{mode}{Get Biosys site information `minimal` (default) or `props`}

\item{type}{List of site identifier}

\item{lat}{Latitude}

\item{long}{Longitude}

\item{distance}{Distance in km from point to search}
}
\value{
Data frame
}
\description{
Get a list of sites around a lat/long point. Optionally include Biosys site
information (such as altitude, slope etc).
}
\examples{
site_info <- get_sites(lat = 51.44553, long = -2.54344, distance = 1, mode = "props")
}
