% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-observations.R
\name{get_observations}
\alias{get_observations}
\title{Get Observations}
\usage{
get_observations(
  site_id = NULL,
  take = 500,
  skip = 0,
  type_id = NULL,
  property_id = NULL,
  ultimate_foi_id = NULL,
  date_from = NULL,
  date_to = NULL
)
}
\arguments{
\item{site_id}{list of sites ids.}

\item{take}{integer max number of returned observations}

\item{skip}{integer number of responses to skip}

\item{type_id}{Observation type}

\item{property_id}{Property type}

\item{ultimate_foi_id}{Ultimate feature of interest}

\item{date_from}{Date string '2020-11-01'}

\item{date_to}{Date string '2021-02-01'}
}
\value{
dataframe
}
\description{
Query observations end point. In the Ecology API, the most central point of
interest are the Observations. Each observation has a single result for a
particular Observable Property for a particular Feature of Interest. This API
accesses hundreds of thousands of surveys and many millions of observations,
this means this endpoint requires at least one filter of site, observable
property, observation type or ultimate feature of interest.
}
\examples{
site_id = c("http://environment.data.gov.uk/ecology/site/bio/43378",
     "http://environment.data.gov.uk/ecology/site/bio/10254")
obs <- get_observations(site_id, take = 25)
}
