% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-site-info.R
\name{get_site_info}
\alias{get_site_info}
\title{Get site info}
\usage{
get_site_info(site_id = NULL)
}
\arguments{
\item{site_id}{Site identifier}
}
\value{
Data frame
}
\description{
Get site info
}
\examples{
site_info <-
get_site_info(site_id = "/environment.data.gov.uk/ecology/site/bio/43378")
}
